Move serializer interfaces to Serializer namespace
