+++ src/Domain/Serializer/NormalizerInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Imdhemy\GooglePlay\Domain;
+namespace Imdhemy\GooglePlay\Domain\Serializer;
 
 interface NormalizerInterface
 {
